Increase `sandbox.waitForStopped` timeout yet again.
For humongous containers (>256GiB of RAM), this can exceed 45s even on
beefy machines.

I considered making the timeout scale based on the memory usage of the
sandbox, but this is only an approximate proxy metric for time to shut down,
and I don't think we can reliably get it once the sandbox process is already
shutting down.

PiperOrigin-RevId: 624009617