runsc: boot: use a more reliable method to check /proc
Currently we check whether /proc is umounted by 'unix.Access'.
If we run runsc as setuid binary using unprivileged user(this is
unsafe, not recommended) we may get EPERM even the /proc is umounted.
This is because access(2) uses the real uid to perform the
permission check but open(2) uses the euid to perform the check.
In setuid cases the uid is not the same as the euid.

In this case we get the error '/proc is stil accessible'.

Notice this error can only occur combined with following two conditions:
1. we uses --network host. This is because in non-host mode, the runsc
will create a new userns and do uid identity mapping, the child process
will be run as 0:0(uid:gid) so the check will be ok
2. the umask setting should be set to mask the all perm of others such as
0027. This is because if we don't mask the others's read perm, the check
will also be ok as the unprivileged uid can also read /proc/ file.

This patch uses a more reliable method to do /proc check.