runsc: add support to control CPU features exposed to user apps
Currently, runsc compares the host feature set between the machine used
for checkpointing and the machine used for restoring. If the former
feature set is not a subset of the latter, the restore will fail, as
user apps might mistakenly use unsupported instructions.

Thanks to cpuid faulting support, it is possible to intercept the cpuid
instruction from user apps and generate results referring to the host
feature set recorded during sandbox boot.

This patch adds support to expose only specified CPU features by
removing features from the list when storing the host feature set. This
makes it possible to checkpoint and restore on machines with different
CPU features. The list of enabled features is passed via annotation.

It should be noted that currently, CPUID faulting is not supported on
the arm64 architecture, and therefore, control over the CPU features
exposed to user apps is also not supported.

Updates #11486

Signed-off-by: Tianyu Zhou <albert.zty@antgroup.com>