Profiling metrics: Support unprefixed data logging.
This is part of a series of changes to add metric charts in performance
benchmarks.

In contexts where profiling metrics are not being intermixed with other
logs, we do not have to prefix the data with a common prefix, nor do we
need the checksum. To keep profiling metric overhead to a minimum in these
cases, this distinguishes two types of writers:

- Normal buffered writers, which do nothing other than buffer writes to
  the underlying writer.
- Line-buffered "lossy" writers, which do the prefixing and checksumming
  that the code was previously doing.

Both can reuse the same underlying write loop (thanks Go generics!).

PiperOrigin-RevId: 631271157