Don't add test jobs if test results are up-of-date in the cache.
Bazel lacks a method to verify if test results are current in the cache.
Additional details can be found at:
https://github.com/bazelbuild/bazel/issues/3978.

To address this limitation, let's change the test runner so that it
doesn't run tests and exits with a non-zero code immediately if the
$TEST_CHECK_UP_TO_DATE_FILE file exists. To check test results in the
cache, we need to create the $TEST_CHECK_UP_TO_DATE_FILE and run tests.
Successful test execution indicates that the results are up-to-date in
the cache.

Signed-off-by: Andrei Vagin <avagin@google.com>