Add `metricsviz` library to process profiling metrics logs.
This is part of a series of changes to add metric charts in performance
benchmarks.

This adds a new library that tests and benchmarks can use to process
profiling metrics logs as generated by the gVisor metrics library.

This current change only implements the parsing part of this library.
A future change will add charting capability.

This lives in a separate package from the pure `metrics` package because
it can be marked test-only as a BUILD target, and because it will soon need
to include a charting library which we do not want to link from the `runsc`
binary.

PiperOrigin-RevId: 631473916