`metricsviz`: Generate charts and HTML webpage of metrics profiling data.
This is part of a series of changes to add metric charts in performance
benchmarks.

This change adds the code to compile the metrics profiling data into charts,
rendering those charts to HTML, and writing this HTML file to a temporary
location.

It uses the Apache echarts library to generate the charts.

PiperOrigin-RevId: 631974672