`metricsviz`: Group relevant metrics on the same chart if applicable.
This is part of a series of changes to add metric charts in performance
benchmarks.

This change creates a list of named sets of metrics ("metric groups").
When a profiling metrics log contains at least two metrics from any given
group, a single chart for that group is created, rather than putting each
of these metrics on their own chart.

Metrics may be part of multiple groups.

If a metric is not in any given group, or if it is the only metric that was
profiled out of the metrics in the group that metric is in, it is charted
individually as per before this change.

PiperOrigin-RevId: 631986612