Add `metricsviz_cli` tool to manually create charts from profiling metrics.
This is part of a series of changes to add metric charts in performance
benchmarks.

Comes with a test, which also serves as an end-to-end-ish test of profiling
metric functionality from collection to charting.

Including `metricsviz` in a `go_binary` requires making it non-`testonly`,
which in turn means removing the `dockerutil` dependency, so this uses an
interface to break the dependency on `dockerutil.Container`.

PiperOrigin-RevId: 633394898