Add timepoints pkg for generating perf metrics via aggregated timestamps
This initial implementation focuses on getting the basics in place, namely:
  - Capturing timestamps that are identifiable by operation instance.
  - Generating metrics from averaging timestamp deltas between every step
    of an operation.

The primary purpose of instrumenting code with these is to provide a
framework for automated performance regression testing. The metrics
may also be interesting data during development to complement pprof
traces.

Further refinements include: improving the API to provide tighter coupling
between Operation-to-TimePoints; and improving performance hit of using
timepoints, possibly making all hotpath functions `nosplit` so that
timepoints could be used in `nosplit` functions.

PiperOrigin-RevId: 640625588