Use blocking sockets when testing raw IP sockets
The original code instantiated the raw IP sockets with NON_BLOCKING,
which caused problems in the PingSuccessfully test case. That case sends
and expects to immediately receive an ICMPv6 Echo Request/Reply. On
Fuchsia, the Loopback devices tx/rx queues are decoupled, so it's
possible for the test to call `recvmsg` before the packet has been
delivered to the socket (resulting in EAGAIN). By using a blocking
socket, the test will wait until `recvmsg` no longer returns EAGAIN.

PiperOrigin-RevId: 641989486