netstack: panic early with TCP state for debugging
We've seen panics in `segment.payloadSize()` stemming from what seems to be an
out of sync write list and sender state. That is, SndUna and SndNxt indicate
there are unacknowledged bytes remaining, but the write list -- which contains
all sent bytes until they're acknowledged -- is empty.

Panic early with TCP state for debugability.

PiperOrigin-RevId: 643120793