test: deflake runsc/boot:boot_test
Due to Go's "zero value" semantics, raw FD values not explicitly specified in
boot.Args default to 0 (stdin). In loader_test.go:TestHostnetWithRawSockets,
which expects a call to boot.New() to fail, this causes the construction of a
fd.FD for Args.ExecFD = 0 which is later closed unpredictably by finalizer.

PiperOrigin-RevId: 644199565