Update gVisor.dev compatibility page and introduction blurb.
This is per user feedback and aims to address common concerns.
The compatibility page was written back when gVisor did not support a
wide variety of common workloads, which is no longer the case; it is now
easier to list what **doesn't** work rather than what **does** work.

Therefore, this change removes the list of supported programs and replaces
it with an explanation of how gVisor ensures that it supports a wide variety
of workloads, and market proofs that this is the case in practice.

PiperOrigin-RevId: 647834933