When starting metric profiling, truncate existing profile log if applicable.
Prior to this change, if `runsc` was configured to use a specific
`--profiling-metrics-log` path, then the second run of this runtime would
overwrite existing metric data in-place, which produces a corrupt file.

This change truncates the existing log, so that the resulting log actually
correspond to what just ran.

This only happens when the `os.File` supports `truncate(2)`, which stdout
does not, so the error is ignored if there is one. It's a best-effort thing.

PiperOrigin-RevId: 648493697