Metrics visualization: Link all charts' X axis together.
This makes it such that hovering over the X axis on one chart shows the same
point on the other charts, and that zooming in on one chart zooms in on all
the others as well.

Sadly this setting doesn't seem to be exposed in the Go bindings of echarts,
so this works by injecting JavaScript code that sets the linkage.

PiperOrigin-RevId: 648519357