Add (probably broken) BuildKite pipeline for refreshing PGO profiles.
This pipeline is intended to run regularly and check if the PGO profiles
need to be refreshed. If they do, it creates a new commit and pushes it.

This will likely not work because the BuildKite agent lacks credentials to
make commits like this, but I would like to submit this anyway and see what
error it gets first. The PGO pipeline is non-blocking and manually
triggered only, so this will not impact other commits from submission.

PiperOrigin-RevId: 648891643