Reword "Supported GPUs" documentation to explain architecture similarities.
This notes which GPU architecture each of the supported GPU types belong to,
and some other consumer-oriented GPUs that also belong to these architectures.
If a workload is broken on them, it is likely also broken on a supported GPU,
and therefore is a valid bug report to improve compatibility of supported GPU
types in gVisor.

It also mentions the existence of the `--nvproxy-driver-version` flag and the
`ioctl_sniffer` tool.

Fixes #10624