Fix sniffer_test to work.
Previously, run_sniffer was not correctly reporting when unsupported ioctls
were found with the compatibility flag set. At the same time, the sniffer test
was not correctly testing a supported CUDA program, since it was using
run_sample which is currently broken with the sniffer.

This also adds the sniffer test to the list of gpu tests.

PiperOrigin-RevId: 663911778