Run post-restore hook in the background.
After restorer.restoreDone() returns successfully, the sandbox has been started
and the application is running. The caller of restorer.restore() is the
"runsc restore" command. Do not hang the "restore" command even after the
sandbox has been restored.

Do the post-restore work in a separate goroutine. If it fails, log a warning
and kill the sandbox.

PiperOrigin-RevId: 674366811