Create runsc chroot directory in a sub-directory of /tmp.
Earlier runsc was using /tmp as the sandbox chroot. However, as indicated in
#10965, some sandbox launches could fail with EBUSY while trying to remount the
/tmp chroot as read-only. We suspect that this could be due to the Golang
runtime or an imported library having an open FD within /tmp.

So we first create a new tmpfs mount over /tmp (like we did before). Then we
additionally create a subdirectory within /tmp and create a second tmpfs mount
there. This subdirectory is used as the chroot. This should protect against the
hypothesized failure scenarios.

Updates #10965

PiperOrigin-RevId: 680652316