Cleanup dev gofer connection if gofer exits but container is still running.
This is helpful in a scenario where:
- The container application is a GPU application.
- The container application exits. Its mount namespace is released, and as a
  result all gofer mounts are released. So all gofer connections are closed,
  except the device gofer connection which is not associated with any mount.
- The gofer process continues to hang waiting for the device gofer connection
  to exit.

Before this change, the device gofer connection was only cleaned upon container
destruction (e.g. via `runsc delete`). After this change, the gofer process is
cleaned up without an explicit `runsc delete`. The gofer monitor is tracking
the rootfs mount's gofer connection. When the application exits, it is notified
that the connection is closed. We can use this signal to clean up the device
gofer connection too.

Updates #10997

PiperOrigin-RevId: 683307750