Fix TestCheckpointRestoreListeningConnection flaky test.
Thee test has a connected connection which is closed (on both sides) before
calling the checkpoint. But sometimes the checkpoint can start before server
side connection is fully closed and cause panic. To fix this, added a sleep
which ensures that the server side connection is fully closed before the
checkpoint.

PiperOrigin-RevId: 685762954