Avoid sending a signal to the container during checkpointing
When checkpointing, the container.SignalContainer will be blocked until
the end of the checkpointing process, during which it cannot acquire the
kernel.extMu mutex. As a result, some runsc commands, such as the runsc
list, will have to wait. The waiting time may vary with the duration of
the checkpoint. This patch adds a flag to prevent sending signals to the
container during checkpointing.

Signed-off-by: Dawei Shen <shendawei.sdw@antgroup.com>