goferfs: Return EOPNOTSUPP when attempting to change mode for symlinks.
As of writing, gVisor does not support fchmodat2(2). Without it, there is no
way to change permissions for a symlink file. Note that fchmod(2) on an O_PATH
FD fails with EBADF. Also fchmodat(2) (which is supported) does not take any
flags and AT_SYMLINK_NOFOLLOW is needed to change permissions on a symlink.

But if in the future we were to support fchmodat2(2), goferfs would return
EBADF because it would try to change mode using fchmod(2) on a O_PATH FD.
Updated goferfs to explicitly return EOPNOTSUPP (consistent with Linux).

Some other minor fixes:
- The comment about why sockets needed special handling for changing mode was
  incorrect. Sockets too have a O_PATH control FD so fchmod(2) on it would fail
  with EBADF. This was the real reason.
- The comment about AT_SYMLINK_NOFOLLOW not being supported is stale. Support
  for fchmodat2(2) along with AT_SYMLINK_NOFOLLOW and AT_EMPTY_PATH flags was
  added in Linux 6.6. Removed the comment for now. We should probably use
  fchmodat2(2) when it is available because it is safer.
- When --directfs=true and the application tries to change mode of a mountpoint
  socket, gofer client makes an RPC to the fsgofer server to do that work.
  fsgofer attempts the mode change using fchmodat(2). However, the fsgofer's
  seccomp filters did not allow fchmodat(2) in --directfs=true mode. Added
  fchmodat(2) to fsgofer seccomp filters unconditionally.

Updates #10385

PiperOrigin-RevId: 686956700