goferfs: Use lgetxattr(2) to get xattr for sockets and symlinks.
In goferfs, the control FD (in gofer client in directfs mode) and host FD in
fsgofer server is an O_PATH FD for socket and symlink files.

fgetxattr(2) fails with EBADF for O_PATH fds. So use lgetxattr(2) instead. This
is a path-based syscall, so it should work for symlinks and sockets. Since the
gofer client can not make path-based syscalls, it falls back to lisafs.

Fixes #11049
Updates #10385

PiperOrigin-RevId: 687160973