gofer: open volumes from the initial userns
when gofer can't open them from its user namespace. The gofer namespace
can have custom user/group mappings that can prevent it to access
volumes.

This change repeats the behaviour of runc. If we can't open a volume from
the gofer userns, we are trying to open it from the initial userns. For
that, we create a small rpc server that the gofer communicates with.

Signed-off-by: Andrei Vagin <avagin@google.com>