`test/kubernetes/testcluster`: Fix `WaitForDaemonset` timeout case.
Prior to this change, `WaitForDaemonset` would erroneously return a
`nil` error when the context expired.

Also make error messages in `WaitForDaemonset` more helpful.
This includes bailing out of the watch-based loop slightly before the
context expiry, and fall back to the poll-based loop which provides a
more helpful error message.

PiperOrigin-RevId: 700805708