`testcluster`: Add support for request retries and recreating k8s clients.
Google's internal test infrastructure is flaky enough that this seems
necessary.

This doesn't wrap all of `kubernetes.Interface` because that is a huge
interface and it would be difficult to replicate it all. Instead, it wraps
all `kubernetes.Interface`-using functions inside a wrapper function where
the passed-in `kubernetes.Interface` object may change across calls.

For regular use, it is still possible to use a simple `kubernetes.Interface`
without further abstraction.

PiperOrigin-RevId: 704502883