Create gofer filestores in gofer's mount namespace.
Before this change, the gofer filestore FDs were opened in the host's mount
namespace. This prevented containerd from unmounting a subcontainer's rootfs
in case (due to a ref leak or bug) the container's rootfs was not destroyed in
gVisor's VFS layer. In such a scenario, the sentry would leak the rootfs
filestore FD and containerd's umount attempts will fail with EBUSY.

This change ensures that all gofer filestores are created in the gofer's mount
namespace, so any unmount attempt from the host mount namespace is not blocked.

This change makes the following 2 logical changes:
1. It introduces a new gofer process synchronization flag --sync-chroot-fd. The
   gofer waits on this FD and then proceeds to set up container filesystem and
   pivot_root(2) + chroot(2) the gofer mount namespace. This is useful if an
   external process wants to take action in gofer's mount namespace while
   having access to the host filesystem. As a result, the existing
   --sync-nvproxy-fd became redundant because it has the same use case. IOW,
   it was just renamed to "sync-chroot-fd" and its meaning became broader.
2. Container.createGoferFilestores() earlier was doing two things: creating
   gofer mount configs and creating all filestores. But the intent of this
   change is to only create filestore FDs after gofer starts in gofer mountns.
   However, the gofer mount configs are needed before that. So this function
   was split into Container.initGoferConfs().

We open FDs from the gofer mount namespace using a neat trick suggested by
Andrei, we prefix the absolute paths with /proc/<gofer-pid>/root. See
proc_pid_root(5) for more details about this.

There are no other intended changes in behavior.

Fixes #9834
Suggested-by: Andrei Vagin <avagin@google.com>
PiperOrigin-RevId: 706880500