Allow IPv6 addresses to be enabled/disabled in runsc based on the sysctl.
By default in runsc, we have both IPv4 and IPv6 addresses enabled on all the
interfaces. However, in runc this is based on the sysctl
net.ipv6.conf.all.disable_ipv6. This CL will make runsc behave similar to runc.
- Implement net.ipv6.conf.all.disable_ipv6 sysctl in runsc. This value will be
copied from the host. By default,the value will be zero.
- If net.ipv6.conf.all.disable_ipv6 is > 0, then only IPv4 addresses will be
enabled on the interfaces including loopback when network mode is "sandbox".
- Added syscall tests to verify the sysctl works correctly.

PiperOrigin-RevId: 733433428