Move nvproxy.objsMu to rootClient.
The objsMu sychronizes 2 things:
- nvproxy object book-keeping
- host ioctl(2) invocations

Since this mutex was on nvproxy device level, nvproxy was synchronizing host
ioctl invocations across clients even though the NVIDIA kernel driver allows
ioctl commands across clients to proceed in parallel.

Move the lock to rootClient and implement locking on a per-client basis. This
is similar to what the NVIDIA kernel driver does; see usages of
_serverLockClientWithLockInfo() in the driver.

As far as we can tell, CUDA programs issue one command to each client at a time
so synchronizing on a per-client basis shouldn't hurt concurrency.

Investigated-by: Jamie Liu <jamieliu@google.com>
Suggested-by: Jamie Liu <jamieliu@google.com>
PiperOrigin-RevId: 735857996