Read pages from the main MemoryFile as early as possible during restore
This change modifies the sandbox restore process such that if the memory
pages are stored as separate files, their contents are optimistically read
from these files as soon as possible during the sandbox restore process,
as opposed to the previous behavior where they were read only after the last
container of the sandbox had been restored.

When memory page data is stored as separate files, the process of restoring
multiple `MemoryFile`s is a bit more intertwined than it appears, for two
reasons:

- The pages metadata file stores data for all `MemoryFile`s one after the
  other. So we must also plumb the pages metadata file reader that is used
  for reading the main `MemoryFile` metadata to the code that also loads the
  private `MemoryFile`s, and ensure that these operations are correctly
  serialized, since the primary `MemoryFile` metadata comes earlier in the
  file.
- The `MemoryFile` page contents are all in the same file, and the byte offset
  within this file at which each each `MemoryFile` begins must be known before
  kicking off async loading of that MemoryFile.

So this change introduces a `PagesFileLoader`, which is meant to be a
thread-safe struct that keeps track of this process. It owns the FDs for the
pages and page metadata files and handles the serialization of the operations
executed using these FDs.

Support for the previous checkpoint/restore format (with all data in a single
file) is maintained using a synchronous alternative that implements the same
interface but simply loads everything from the shared state file reader when
the time to do so comes around.

PiperOrigin-RevId: 739347254