Protect Task.ptraceSeized with TaskSet mutex and Signal mutex.
Previously Task.ptraceSeized was protected only by TaskSet mutex, but some
pathways incorrectly modified the field without holding the lock, leading to
data races.

This CL changes the locking so that TaskSet and Signal mutexes are required to
write Task.ptraceSeized, but only one of them must be held to read
Task.ptraceSeized.

Reported-by: syzbot+b0de1abe368849ca2d00@syzkaller.appspotmail.com
PiperOrigin-RevId: 742317413