systrap: Initialize stub threads in the same way on both x86 and ARM.
Initializing threads via a custom Systrap entrypoint (as was done on x86 prior
to this CL) is incompatible with debuggers, because they set the single-step
flag. If a brand new stub thread managed to pick up an already existing context
with this flag set, it would immediately invoke the sighandler upon restoring
RFLAGS, which would corrupt the stack and instruction pointers.

PiperOrigin-RevId: 761250233