Allow sched_getaffinity in syscall filters
Go 1.25's automatic GOMAXPROCS background updates will periodically call
sched_getaffinity (for total CPU count) and pread64 (for cgroup quota limit).
The latter is already allowed by the filters, but the former is not.

In the sentry, the explicit runtime.GOMAXPROCS call at startup disables the
runtime's automatic updates. In theory this makes the filter unnecessary,
however the runtime only actually guarantees it won't change the value of
GOMAXPROCS after runtime.GOMAXPROCS. It does not guarantee that a concurrent
update run won't call the syscalls after runtime.GOMAXPROCS returns (when this
happens, the runtime by definition must later discard any change it finds).

That means it is theoretically possible for a background sched_getaffinity call
to occur after filters are installed.

This lack of guarantee makes the feature difficult to work with, so I intend to
change the runtime to provide a stronger guarantee, but until then I don't
think it hurts to allow this system call.

I haven't actually seen a failure due to a concurrent update yet, this is
precautionary.

Note that the gofer does not explicitly set GOMAXPROCS, so it will continue to
need the filter unless that changes.

PiperOrigin-RevId: 764381364