overlayfs: Do not SetStat() UID/GID for newly created file.
Instead, use a credentials object which has the target UID/GID set as the
EffectiveKUID/EffectiveKGID. This will ensure that the file is created with the
right owners. This is consistent with what Linux does. This avoids making
unnecessary SetStatAt() calls to the VFS, which may incur additional RPCs.

This commit additionally fixes the following bugs:
- When parent directory has the SGID bit set, newly created symlinks should
  inherit the GID from the parent directory. This was not happening earlier.
- There was a bug in newChildOwnerStat() where it was checking an uninitialized
  `stat.Mode` to check if the newly created file is a directory or not. That
  would always evaluate to false, so child directories would not have the SGID
  bit set.

PiperOrigin-RevId: 768239284