gofer: Support restore of deleted files whose original path is occupied.
It is possible that the application can open an FD to a file, delete the file,
create another file at that path and then checkpoint the container. In such a
scenario, our gofer restore logic will fail with EEXIST because a file already
exists on the host at that location when we try to recreate it.

Use a temporary filename to recreate the file in such a scenario. This file is
deleted again anyways.

Fixes #11425

PiperOrigin-RevId: 768263262