Add initial VM qualification suite
Add Makefile target with the initial subset of GPU tests that qualify
new VMs. The other tests in gpu-all-tests don't work on newer models
and may or may not be fixed.

This give a good idea of coverage of the existing tests so that more
can be developed, or we can fix the gaps in existing tests.

PiperOrigin-RevId: 769777107