PGO update script: Connect to GitHub over HTTPS rather than SSH.
The `gh auth setup-git` command appears to configure Git as a credential
helper for `https`, not for SSH-based access. SSH access fails with the
host authenticity check.

Try to connect over HTTPS instead, to see if that works properly.
Otherwise we will have to provision `~/.ssh/known_hosts` and give the
agent a private SSH key for authentication.

PiperOrigin-RevId: 772156228