state: Maintain a separate list with leaf nodes for objectDecodeState.
Profiling shows that we spend a LOT of time in decodeState.Load() in firing all
callbacks in the right order. objectDecodeState is organized as a DAG.
Callbacks need to be fired on leaf nodes (where blockedBy==0) and then fired
on internal nodes only when all their dependencies have executed their
callbacks. The earlier approach was to continuously scan a HUGE list of pending
nodes searching for a leaf node, fire callbacks on that, check if any of its
dependents became leaves and then scan the full list again. We were wasting a
lot of time re-evaluating internal nodes repeatedly. For instance, the kernel
object created by checkpointing SGLang server running "gemma-3-27b-pt" had
368,245 objects. The whole pending list was scanned 238,102 times. We spent
2 minutes 10 seconds in just executing callbacks. The total kernel load time
was 2 minutes 30 seconds.

In this change, we track leaf nodes in a separate linked list. We keep
appending to it as and when new nodes become leaves. This way we don't need to
search for leaves. This improved kernel load time to 12 seconds (-92%).

PiperOrigin-RevId: 775402687