Rename WARN* functions to BugTraceback*, and optimize them for inlining.
The rename is todo with following a more Go-style naming convention. To this
end, the "default" BugTraceback will be the one that takes a single error
parameter.

As to inlining, the formatting variant required pulling the Sprintf call into
the helper, and the BugTraceback(err) variant seems to require a hack from
pkg/sync/gate_unsafe.go:Enter where there is a noinline helper.

PiperOrigin-RevId: 789915634