runsc: Enable cpu-num-from-quota by default
By enabling this feature, the Sentry will report a more accurate number of CPUs
to the guest. This can be important for performance, as reporting an incorrect
CPU count can lead to either CPU overcommit or CPU undercommit. Both of these
conditions can negatively affect application performance.

The logic for determining guest CPUs is as follows:
* If the container spec has no cgroup configuration, runsc uses the number
  reported by runtime.NumCPU(). The sentry is running in an isolated mount
  namespace, so it doesn't have access to host cgroups.

* If there is a cgroup configuration, runsc calculates the number of CPUs based
  on the cgroup cpuset.cpus. If cpu-num-from-quota is set, in addition to
  cpuset.cpus, it calculate a number of CPUs based on the specified cgroup CPU
  quota and uses the minimum of these two values.

While there are some known caveats with CPU quotas (as described in
golang/go#73193), this method provides a better default experience for most
users by more accurately reflecting the resources allocated to the container.
For users who need to override this behavior, the cpu-num-from-quota option
can still be explicitly set to false.

The pull request that introduced the cpu-num-from-quota option initially
suggested setting its default value to false to align with Docker's behavior.
This decision has been revised here, primarily because the host's cgroup
configuration is not exposed to the sandbox. Additionally, this change better
supports applications that are not cgroup-aware.

Here is one real-life example where this change has an effect. Golang
applications compiled with go1.25 set GOMAXPROCS according to cgroup
CPU limits:
```
$ docker run --cpus 4 -v /tmp/nproc:/tmp/nproc -it --rm alpine \
    /tmp/nproc/main
GOMAXPROCS 4 NumCPU 96
```
Without this change, in gVisor, GOMAXPROCS is set incorrectly:
```
$ docker run --cpus 4 -v /tmp/nproc:/tmp/nproc -it --rm \
    --runtime runsc alpine /tmp/nproc/main
GOMAXPROCS 96 NumCPU 96
```
With this change, GOMAXPROCS is set correctly:
```
$ docker run --cpus 4 -v /tmp/nproc:/tmp/nproc -it --rm \
    --runtime runsc-cpu-quota alpine /tmp/nproc/main
GOMAXPROCS 4 NumCPU 4
```

PiperOrigin-RevId: 789970027