Update protocol to parse fields from network order to host order.
Some fields in the NETFILTER protocol are expected to be received in
network order. This change updates both the NETFILTER protocol to
do this conversion for those fields, as well as updating the tests
to send those fields in network byte order.

The change also ensures that the tests wait for the correct number
of response messages from the kernel when reading from the NETFILTER
sockets.

Updates #11778

PiperOrigin-RevId: 792337707