pkg/sentry: add cleanup to avoid refcount leak
Currently, on certain error paths during task creation (e.g., when
"runsc exec" fails), we do not properly handle the refcount of the
mntns/FSContext, leading to a refcount leak. This patch adds cleanup
that would be released before the reference ownership is transferred
to callee, ensuring that the refcount is correctly decremented.

Fixes #12054

Signed-off-by: Tianyu Zhou <albert.zty@antgroup.com>