PGO: Run benchmarks for a shorter time, but force GC to run more often.
The PGO profiles don't need to run that long, but they do need to ensure
GC happens a few times to exercise GC-triggered codepaths (e.g. on
`sync.Pool` objects, network packet chunk pools, etc).

This change introduces a flag to force GC to run on a set interval during
CPU profiling, and enables it for PGO benchmark runs. It also reduces the
PGO benchmark time target, since GC is now guaranteed to run a couple of
times.

It also deletes all profiles, in order to force them to be refreshed
with this change in effect.

PiperOrigin-RevId: 797835198