Allow IP_(ADD|REMOVE)_MEMBERSHIP on AF_INET6 sockets
Just like Linux does.

The new test `AddV4MembershipToV6Socket` passes only if there is a routing
table entry to resolve the ipv4 multicast address, and that is easiest to setup
via a "external networking" test. Now it turns out that the ipv6 "external
networking" test did not actually have a second non-loopback (external)
interface, so I've added that.

PiperOrigin-RevId: 807959283