gofer: Invalidate dentry if file type changes but inode number doesn't.
For some filesystems like ext4, it possible that the inode number is re-used in
between unlink and file creation. Hence it is possible a file is replaced on
the host filesystem with another file but the inode number remains the same.
The gofer client can't detect such changes and trigger invalidation. But at
least we can detect if the file type has changed and invalidate the old file.

PiperOrigin-RevId: 810536191