Consolidate metadata configuration during S/R
S/R requires certain metadata objects to be set to the state file.
The code doing that was spreadout, duplicating some of the logic
(spec) and missing metadata in others (container count).

This change, moves the code to a single location that is called from
the multiple places that trigger saving.

Fixes #1956

PiperOrigin-RevId: 811463154