Handle restore failure more robustly.
- This change most notably adds a new loaderState "restoreFailed". As of
  writing, this is used to classify restore failures which occur after the last
  container's restore command is received. See future work section.
- `runsc wait --restore` now returns the restore error, if one occurred.
- Added more explicit checks for all values of loaderState and error out when
  an unexpected state is reached.
- `restorer` has been simplified to save a pointer to `cm` instead of using
  callback functions, which are harder to follow (and step-into).

Future work:
We still do not handle failures that occur when intermediate container's
restore command fails. Or if we receive an unexpected "start" command instead
of "restore". In such scenarios, it is unlikely that we will recover to have
a successful restore as it suggests some issue with the higher level tooling
issuing these commands.

PiperOrigin-RevId: 831757593