Make --allow-packet-socket-write a public flag.
This flag is now required for Docker versions 28 and beyond to function
inside gVisor, as dockerd sends unsolicited ARP/NA requests using AF_PACKET
sockets. The flag is no longer test-only.

The docker behavior to send the ARP/NA pkts was introduced in eaa84bc [1].
Failure to send was made a hard error in 422e056 [2], and in bc130f36 [3],
docker evolved to rely on these packets being sent.

[1] https://github.com/moby/moby/commit/eaa84bc8f4
[2] https://github.com/moby/moby/commit/422e056b0f
[3] https://github.com/moby/moby/commit/bc130f367d

PiperOrigin-RevId: 840805091