Account for endpoint header size when calculating Route MTU.
If r.mtu has been set, as it is after
https://github.com/google/gvisor/pull/11857, we skip calculating the proper MTU
via endpoint.MTU and use r.mtu instead. However nothing in the callstack from
tcpip/stack/stack.go:FindRoute to setting r.mtu actually adjusts the size to be
able to hold the endpoint header. Hence we need to adjust it.

PiperOrigin-RevId: 843416148