Fix execveat pathname construction when called on memfd.
In execveat, when dirfd is not AT_FDCWD and pathname is not absolute, the
pathname is now constructed using /dev/fd/. Earlier, we were using MappedName()
which could return paths with " (deleted)" suffix for deleted files, making
them inaccessible from filesystem traversal.

This is consistent with that Linux does, see fs/exec.c:alloc_bprm().

Suggested-by: @dean-deng

Fixes #1090

PiperOrigin-RevId: 849106927